fixup! fix(DOM): iOS safari pinch zooming fix
